+++ 2.3/rtb/native.go
@@ -41,7 +41,7 @@ type Native struct {
 	// Description:
 	//   List of supported API frameworks for this impression. Refer to
 	//   List 5.6. If an API is not explicitly listed, it is assumed not to be supported.
-	API []int64 `json:"api"`
+	API []uint8 `json:"api"`
 
 	// Attribute:
 	//   sequence
@@ -49,7 +49,7 @@ type Native struct {
 	//   integer array
 	// Description:
 	//   Blocked creative attributes. Refer to List 5.3.
-	BAttr []int64 `json:"battr"`
+	BAttr []uint8 `json:"battr"`
 
 	// Attribute:
 	//   ext
